refactor: gir parts page data fremfor augmented
